feat(DTFS2-7217): commonise FED endpoints
