+++ gef-ui/api-tests/unisued-facilties.api-test.js
@@ -1,7 +1,7 @@
-const { MAKER } = require('../../server/constants/roles');
-const { withRoleValidationApiTests } = require('../common-tests/role-validation-api-tests');
-const app = require('../../server/createApp');
-const { get, post } = require('../create-api').createApi(app);
+const { MAKER } = require('../server/constants/roles');
+const { withRoleValidationApiTests } = require('./common-tests/role-validation-api-tests');
+const app = require('../server/createApp');
+const { get, post } = require('./create-api').createApi(app);
 
 const dealId = '123';
 const facilityId = '111';
