+++ gef-ui/server/controllers/facility-end-date/index.ts
@@ -3,22 +3,52 @@ import { getDate, getMonth, getYear, isSameDay, parseISO } from 'date-fns';
 import { CustomExpressRequest, isFacilityEndDateEnabledOnGefVersion, parseDealVersion } from '@ukef/dtfs2-common';
 import { isTrueSet, validationErrorHandler } from '../../utils/helpers';
 import * as api from '../../services/api';
-import { validateAndParseFacilityEndDate } from '../../utils/validate-facility-end-date';
-import { asLoggedInUserSession } from '../../utils/express-session';
+import { validateAndParseFacilityEndDate } from './validation';
+import { asLoggedInUserSession, LoggedInUserSession } from '../../utils/express-session';
 import { FacilityEndDateViewModel } from '../../types/view-models/facility-end-date-view-model';
 import { getCoverStartDateOrStartOfToday } from '../../utils/get-cover-start-date-or-start-of-today';
 
 type FacilityEndDateParams = { dealId: string; facilityId: string };
 type FacilityEndDatePostBody = { 'facility-end-date-day': string; 'facility-end-date-month': string; 'facility-end-date-year': string };
 
-type GetFacilityEndDateRequest = CustomExpressRequest<{ params: FacilityEndDateParams; query: { status: string } }>;
+type GetFacilityEndDateRequest = CustomExpressRequest<{ params: FacilityEndDateParams; query: { status: string | undefined } }>;
 type PostFacilityEndDateRequest = CustomExpressRequest<{
   reqBody: FacilityEndDatePostBody;
   params: FacilityEndDateParams;
-  query: { saveAndReturn: string; status: string };
+  query: { saveAndReturn: string; status: string | undefined };
 }>;
 
-export const getFacilityEndDate = async (req: GetFacilityEndDateRequest, res: Response) => {
+type PostRequestUris = {
+  nextPage: string;
+  previousPage: string;
+  saveAndReturn: string;
+};
+
+const getFacilityEndDateViewModel = (facility: Record<string, unknown>, previousPage: string, status: string | undefined): FacilityEndDateViewModel => {
+  if (typeof facility.dealId !== 'string' || typeof facility._id !== 'string') {
+    throw new Error('Invalid facility ids stored in database');
+  }
+
+  const facilityEndDateViewModel: FacilityEndDateViewModel = {
+    dealId: facility.dealId,
+    facilityId: facility._id,
+    previousPage,
+    status,
+  };
+
+  if (typeof facility.facilityEndDate === 'string') {
+    const facilityEndDate = parseISO(facility.facilityEndDate);
+    facilityEndDateViewModel.facilityEndDate = {
+      day: String(getDate(facilityEndDate)),
+      month: String(getMonth(facilityEndDate) + 1),
+      year: String(getYear(facilityEndDate)),
+    };
+  }
+
+  return facilityEndDateViewModel;
+};
+
+const handleGetFacilityEndDate = async (req: GetFacilityEndDateRequest, res: Response, previousPage: string) => {
   const {
     params: { dealId, facilityId },
     query: { status },
@@ -29,34 +59,54 @@ export const getFacilityEndDate = async (req: GetFacilityEndDateRequest, res: Re
     const { details: facility } = (await api.getFacility({ facilityId, userToken })) as { details: Record<string, unknown> };
     const deal = (await api.getApplication({ dealId, userToken })) as Record<string, unknown> & { version?: number };
 
-    if (!isFacilityEndDateEnabledOnGefVersion(parseDealVersion(deal.version)) || !facility.isUsingFacilityEndDate) {
-      return res.redirect(`/gef/application-details/${dealId}`);
+    const shouldRedirectFromPage = !isFacilityEndDateEnabledOnGefVersion(parseDealVersion(deal.version)) || !facility.isUsingFacilityEndDate;
+
+    if (shouldRedirectFromPage) {
+      return res.redirect(previousPage);
     }
 
-    const facilityEndDateViewModel: FacilityEndDateViewModel = {
-      dealId,
-      facilityId,
-      status,
-      previousPage: `/gef/application-details/${dealId}/facilities/${facilityId}/about-facility`,
-    };
-
-    if ('facilityEndDate' in facility && typeof facility.facilityEndDate === 'string') {
-      const facilityEndDate = parseISO(facility.facilityEndDate);
-      facilityEndDateViewModel.facilityEndDate = {
-        day: String(getDate(facilityEndDate)),
-        month: String(getMonth(facilityEndDate) + 1),
-        year: String(getYear(facilityEndDate)),
-      };
-    }
-
-    return res.render('partials/facility-end-date.njk', facilityEndDateViewModel);
+    return res.render('partials/facility-end-date.njk', getFacilityEndDateViewModel(facility, previousPage, status));
   } catch (error) {
     console.error(error);
     return res.render('partials/problem-with-service.njk');
   }
 };
 
-export const postFacilityEndDate = async (req: PostFacilityEndDateRequest, res: Response) => {
+export const getFacilityEndDateFromUnissuedFacilitiesPage = async (req: GetFacilityEndDateRequest, res: Response) =>
+  handleGetFacilityEndDate(req, res, `/gef/application-details/${req.params.dealId}/unissued-facilities/${req.params.facilityId}/about`);
+
+export const getFacilityEndDateFromApplicationPreviewPage = async (req: GetFacilityEndDateRequest, res: Response) =>
+  handleGetFacilityEndDate(req, res, `/gef/application-details/${req.params.dealId}/unissued-facilities/${req.params.facilityId}/change`);
+
+export const getFacilityEndDateFromApplicationDetailsPage = async (req: GetFacilityEndDateRequest, res: Response) =>
+  handleGetFacilityEndDate(req, res, `/gef/application-details/${req.params.dealId}/facilities/${req.params.facilityId}/about-facility`);
+
+const updateFacilityEndDateIfChanged = async (
+  existingFacility: Record<string, unknown>,
+  facilityEndDate: Date,
+  userSession: LoggedInUserSession,
+): Promise<void> => {
+  const facilityEndDateNeedsUpdating =
+    typeof existingFacility.facilityEndDate !== 'string' || !isSameDay(parseISO(existingFacility.facilityEndDate), facilityEndDate);
+
+  if (!facilityEndDateNeedsUpdating) {
+    return;
+  }
+  await api.updateFacility({
+    facilityId: existingFacility._id,
+    payload: {
+      facilityEndDate,
+    },
+    userToken: userSession.userToken,
+  });
+
+  const applicationUpdate = {
+    editorId: userSession.user._id,
+  };
+  await api.updateApplication({ dealId: existingFacility.dealId, application: applicationUpdate, userToken: userSession.userToken });
+};
+
+const handlePostFacilityEndDate = async (req: PostFacilityEndDateRequest, res: Response, uris: PostRequestUris) => {
   try {
     const {
       params: { dealId, facilityId },
@@ -64,15 +114,16 @@ export const postFacilityEndDate = async (req: PostFacilityEndDateRequest, res: 
       query: { saveAndReturn, status },
     } = req;
 
-    const { userToken, user } = asLoggedInUserSession(req.session);
+    const userSession = asLoggedInUserSession(req.session);
 
     const facilityEndDateIsBlank = !facilityEndDateYear && !facilityEndDateMonth && !facilityEndDateDay;
 
+    // If the user clicks save and return with no values filled in, we do not update the database
     if (isTrueSet(saveAndReturn) && facilityEndDateIsBlank) {
-      return res.redirect(`/gef/application-details/${dealId}`);
+      return res.redirect(uris.saveAndReturn);
     }
 
-    const { details: facility } = (await api.getFacility({ facilityId, userToken })) as { details: Record<string, unknown> };
+    const { details: facility } = (await api.getFacility({ facilityId, userToken: userSession.userToken })) as { details: Record<string, unknown> };
 
     const facilityEndDateErrorsAndDate = validateAndParseFacilityEndDate(
       {
@@ -84,7 +135,7 @@ export const postFacilityEndDate = async (req: PostFacilityEndDateRequest, res: 
     );
 
     if ('errors' in facilityEndDateErrorsAndDate) {
-      return res.render('partials/facility-end-date.njk', {
+      const facilityEndDateViewModel: FacilityEndDateViewModel = {
         errors: validationErrorHandler(facilityEndDateErrorsAndDate.errors),
         dealId,
         facilityId,
@@ -93,36 +144,42 @@ export const postFacilityEndDate = async (req: PostFacilityEndDateRequest, res: 
           month: facilityEndDateMonth,
           year: facilityEndDateYear,
         },
+        previousPage: uris.previousPage,
         status,
-      });
-    }
-
-    const facilityEndDate = facilityEndDateErrorsAndDate.date;
-
-    const facilityEndDateValueIsUnchanged = typeof facility.facilityEndDate === 'string' && isSameDay(parseISO(facility.facilityEndDate), facilityEndDate);
-
-    if (!facilityEndDateValueIsUnchanged) {
-      await api.updateFacility({
-        facilityId,
-        payload: {
-          facilityEndDate,
-        },
-        userToken,
-      });
-
-      const applicationUpdate = {
-        editorId: user._id,
       };
-      await api.updateApplication({ dealId, application: applicationUpdate, userToken });
+      return res.render('partials/facility-end-date.njk', facilityEndDateViewModel);
     }
+
+    await updateFacilityEndDateIfChanged(facility, facilityEndDateErrorsAndDate.date, asLoggedInUserSession(req.session));
 
     if (isTrueSet(saveAndReturn)) {
-      return res.redirect(`/gef/application-details/${dealId}`);
+      return res.redirect(uris.saveAndReturn);
     }
 
-    return res.redirect(`/gef/application-details/${dealId}/facilities/${facilityId}/provided-facility`);
+    return res.redirect(uris.nextPage);
   } catch (error) {
     console.error(error);
     return res.render('partials/problem-with-service.njk');
   }
 };
+
+export const postFacilityEndDateFromApplicationPreviewPage = async (req: PostFacilityEndDateRequest, res: Response) =>
+  handlePostFacilityEndDate(req, res, {
+    nextPage: `/gef/application-details/${req.params.dealId}`,
+    saveAndReturn: `/gef/application-details/${req.params.dealId}`,
+    previousPage: `/gef/application-details/${req.params.dealId}/unissued-facilities/${req.params.facilityId}/change`,
+  });
+
+export const postFacilityEndDateFromUnissuedFacilitiesPage = async (req: PostFacilityEndDateRequest, res: Response) =>
+  handlePostFacilityEndDate(req, res, {
+    nextPage: `/gef/application-details/${req.params.dealId}/unissued-facilities`,
+    saveAndReturn: `/gef/application-details/${req.params.dealId}/unissued-facilities`,
+    previousPage: `/gef/application-details/${req.params.dealId}/unissued-facilities/${req.params.facilityId}/about`,
+  });
+
+export const postFacilityEndDateFromApplicationDetailsPage = async (req: PostFacilityEndDateRequest, res: Response) =>
+  handlePostFacilityEndDate(req, res, {
+    nextPage: `/gef/application-details/${req.params.dealId}/facilities/${req.params.facilityId}/provided-facility`,
+    saveAndReturn: `/gef/application-details/${req.params.dealId}`,
+    previousPage: `/gef/application-details/${req.params.dealId}/facilities/${req.params.facilityId}/about-facility`,
+  });
