+++ gef-ui/server/controllers/facility-end-date/validation.test.ts
@@ -1,8 +1,8 @@
 import { add } from 'date-fns';
-import { validateAndParseDayMonthYear } from './day-month-year-validation';
-import { validateAndParseFacilityEndDate } from './validate-facility-end-date';
+import { validateAndParseDayMonthYear } from '../../utils/day-month-year-validation';
+import { validateAndParseFacilityEndDate } from './validation';
 
-jest.mock('./day-month-year-validation');
+jest.mock('../../utils/day-month-year-validation');
 
 const mockValidateAndParseDayMonthYear = validateAndParseDayMonthYear as unknown as jest.Mock;
