+++ gef-ui/server/controllers/facility-end-date/validation.ts
@@ -1,9 +1,9 @@
 import { add, isAfter, isBefore, startOfDay } from 'date-fns';
 import { uniq } from 'lodash';
 import { FACILITY_END_DATE_MAXIMUM_YEARS_IN_FUTURE } from '@ukef/dtfs2-common';
-import { validateAndParseDayMonthYear } from './day-month-year-validation';
-import { DayMonthYear } from '../types/date';
-import { ErrorsOrDate } from '../types/errors-or-date';
+import { validateAndParseDayMonthYear } from '../../utils/day-month-year-validation';
+import { DayMonthYear } from '../../types/date';
+import { ErrorsOrDate } from '../../types/errors-or-date';
 
 export const validateAndParseFacilityEndDate = (dayMonthYear: DayMonthYear, coverStartDate: Date): ErrorsOrDate => {
   const errRef = 'facilityEndDate';
