+++ gef-ui/server/routes/facilities/facility-end-date.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-misused-promises */
 import express from 'express';
-import { getFacilityEndDate, postFacilityEndDate } from '../../controllers/facility-end-date';
+import { getFacilityEndDateFromApplicationDetailsPage, postFacilityEndDateFromApplicationDetailsPage } from '../../controllers/facility-end-date/index';
 import { validateRole, validateToken, validateBank } from '../../middleware';
 import { MAKER } from '../../constants/roles';
 
@@ -9,7 +9,7 @@ const router = express.Router();
 router
   .route('/application-details/:dealId/facilities/:facilityId/facility-end-date')
   .all([validateToken, validateBank, validateRole({ role: [MAKER] })])
-  .get(getFacilityEndDate)
-  .post(postFacilityEndDate);
+  .get(getFacilityEndDateFromApplicationDetailsPage)
+  .post(postFacilityEndDateFromApplicationDetailsPage);
 
 export default router;
